added lastKeyDown variables to main menu

+++ BrickBreaker/Highscore.cs
@@ -28,7 +28,7 @@ namespace BrickBreaker
 
             for (int i = 0; i < Convert.ToInt16(Form1.highscoreList.Count); i++)
             {
-                if (Form1.currentScore > Convert.ToInt16(Form1.highscoreList[i].score))
+                if (Form1.currentScore > Convert.ToInt32(Form1.highscoreList[i].score))
                 {
                     Form1.highscoreList.Insert(i, hs);
